+++ packages/api/src/defaultStyleOptions.ts
@@ -292,7 +292,9 @@ const DEFAULT_OPTIONS: Required<StyleOptions> = {
   // Video
   videoHeight: 270, // based on bubbleMaxWidth: 480 / 16 * 9 = 270
 
-  maxMessageLength: 2000
+  maxMessageLength: 2000,
+
+  stylesRoot: document.head
 };
 
 export default DEFAULT_OPTIONS;