+++ packages/test/page-object/src/globals/pageElements/index.js
@@ -9,6 +9,7 @@ import focusedActivity from './focusedActivity';
 import linkDefinitions from './linkDefinitions';
 import liveRegion from './liveRegion';
 import microphoneButton from './microphoneButton';
+import root from './root';
 import scrollToEndButton from './scrollToEndButton';
 import sendBoxTextBox from './sendBoxTextBox';
 import sendButton from './sendButton';
@@ -34,6 +35,7 @@ export {
   linkDefinitions,
   liveRegion,
   microphoneButton,
+  root,
   scrollToEndButton,
   sendBoxTextBox,
   sendButton,